requirements.txt
# Core dependencies
qdrant-client[fastembed]>=1.7.1
gradio==4.15.0
groq==0.4.1
pypdf==3.17.4
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
sentence-transformers==2.2.2

# ML and AI dependencies
torch>=2.0.0
transformers>=4.36.0
numpy>=1.24.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Logging and utilities
colorlog>=6.7.0
tqdm>=4.66.0

# HTTP and web dependencies (for Gradio)
requests>=2.31.0
aiofiles>=23.2.1